refactor: Renamed LinkButton to Button :truck:
